pwdlock: Use external program to lock passwd file
This allows a SELinux transition, so that write access to the passwd file is
restricted to a small utility.

While writing to /etc/passwd is not required, SELinux sets it up so that the
.pwd.lock file is in the same context, so full access to it is required here.

The new systemd-pwdlock utility locks the file and passes the open fd through
a unix socket, allowing the system manager to keep the file locked for the
duration it requires.

This also requires a SELinux policy change to create the new systemd_pwdlock_t
type and allow it to perform the required operations.

Tested: using DynamicUser=yes on Fedora with SELinux enabled.