hacking: Document hacking on Fedora using rpmbuild
This method uses `rpmbuild --build-in-place` to share the local build tree with
normal development. Using the Fedora RPM specfile ensures the options used for
development match those used by the official RPMs, so everything that needs to
be enabled will be.

Also add a gitignore for the temporary files created during the rpmbuild.