[WIP] Introduce MemoryAutoPilot= directive (senpai)
This uses the senpai approach to allow measuring accurate memory requirements
of services.

It does so by tuning memory.high in order to apply increasing pressure on a
cgroup, then monitoring memory.pressure to check when the processes are
actually going to start feeling memory pressure, at which point it will backoff
and increase memory.high again, keeping on that cycle to ensure unused cache
pages and buffers actually get reclaimed quickly while ensuring the process
performance is essentially not affected.

It can be used with:

  MemoryAutoPilot=yes

on a service unit, or:

  # systemd-run -p MemoryAutoPilot=yes ...

Or:

  # systemctl set-property xyz.service MemoryAutoPilot=yes

It can also be disabled dynamically.

See https://github.com/facebookincubator/senpai and also Johannes' presentation
about it at All Systems Go! on Youtube: https://youtu.be/ujk2pfgPul8

This implementation still has quite some room for improvement, but I would like
to get some early feedback so I can polish it while already fixing issues we
uncover in discussions about the approach (and name of directive, parameters we
want to allow to tweak, etc.)